scripts/inference.py
"""
Inference script for SRD2026 model
"""

import argparse
import torch
from pathlib import Path
from PIL import Image


def parse_args():
    parser = argparse.ArgumentParser(description="Run inference with SRD2026 model")
    parser.add_argument(
        "--image",
        type=str,
        required=True,
        help="Path to input image"
    )
    parser.add_argument(
        "--checkpoint",
        type=str,
        required=True,
        help="Path to model checkpoint"
    )
    parser.add_argument(
        "--output",
        type=str,
        default=None,
        help="Path to save output (optional)"
    )
    return parser.parse_args()


def main():
    args = parse_args()
    
    print(f"Loading model from: {args.checkpoint}")
    print(f"Processing image: {args.image}")
    
    # TODO: Implement inference
    print("Inference implementation will go here")
    

if __name__ == "__main__":
    main()
