scripts/train.py
"""
Training script for SRD2026 model
"""

import argparse
import yaml
import torch
from pathlib import Path


def parse_args():
    parser = argparse.ArgumentParser(description="Train SRD2026 model")
    parser.add_argument(
        "--config",
        type=str,
        required=True,
        help="Path to configuration file"
    )
    parser.add_argument(
        "--resume",
        type=str,
        default=None,
        help="Path to checkpoint to resume training"
    )
    parser.add_argument(
        "--output_dir",
        type=str,
        default="./checkpoints",
        help="Directory to save checkpoints"
    )
    return parser.parse_args()


def main():
    args = parse_args()
    
    # Load configuration
    with open(args.config, 'r') as f:
        config = yaml.safe_load(f)
    
    print(f"Configuration loaded from: {args.config}")
    print(f"Output directory: {args.output_dir}")
    
    # TODO: Implement training loop
    print("Training implementation will go here")
    

if __name__ == "__main__":
    main()
