requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0

# Multimodal LLM dependencies
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Data processing
opencv-python>=4.7.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Evaluation metrics
pycocoevalcap>=1.2
nltk>=3.8
